+++ techpack/camera/drivers/oplus/cam_sensor_module/cam_ois/onsemi_fw/LC898128/PhoneUpdate.h
@@ -313,7 +313,7 @@ typedef union ULLNVAL	UnllnVal;
 //==============================================================================
 //extern UINT_8	FlashDownload128( UINT_8 , UINT_8, UINT_8  );
 
-extern UINT_8	SetAngleCorrection( float , UINT_8 , UINT_8  );
+extern UINT_8	SetAngleCorrection( long , UINT_8 , UINT_8  );
 extern UINT_8	UnlockCodeSet( void );
 extern UINT_8	UnlockCodeClear(void);
 extern UINT_32	MeasGyAcOffset(  void  );
